+++ src/main.ts
@@ -35,7 +35,7 @@ import {Stats} from './stats/stats';
 import profiler from './profiler/screeps-profiler';
 import _Overmind from './Overmind_obfuscated'; // this should be './Overmind_obfuscated' unless you are me
 import {VersionMigration} from './versionMigration/migrator';
-import {RemoteDebugger} from './console/remoteDebugger';
+import {RemoteDebugger} from './debug/remoteDebugger';
 // =====================================================================================================================
 
 // @formatter:on