+++ src/overlords/core/upgrader.ts
@@ -74,7 +74,7 @@ export class UpgradingOverlord extends Overlord {
 				upgrader.task = Tasks.withdraw(this.upgradeSite.battery);
 			}
 			// Find somewhere else to recharge from
-			else {
+			else { // TODO: BUG HERE IF NO UPGRADE CONTAINER
 				if (this.upgradeSite.battery && this.upgradeSite.battery.targetedBy.length == 0) {
 					upgrader.task = Tasks.recharge();
 				}