+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file. The format 
 ## [Unreleased]
 
 ### Added
-- Added the `RemoteDebugger` module, which lets me remotely debug other Overmind players' code by communicating through public memory segments. 
+- Added the `RemoteDebugger` module, which lets me remotely debug other Overmind players' code in real-time by communicating through public memory segments. 
     - You can start and end a debug session with the `startRemoteDebugSession()` and `endRemoteDebugSession()` commands
     - Debug sessions automatically time out after 1000 ticks unless extended
     - Ping me on Slack #overmind if you want me to debug something for you
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file. The format 
 - Fixed a bug in `WorkerOverlord` where workers would not fortify ramparts to the needed hits to withstand multiple stacked nuclear strikes
 - Fixed a `RoadPlanner` bug which caused it to prever pathing roads along edge tiles between rooms
 - Fixed a recently-introduced bug which prevented drones from repairing their containers
+
 
 
 ## Overmind [0.5.2.1] - 2019.2.8